% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_track_metadata.R
\name{clean_track_metadata}
\alias{clean_track_metadata}
\title{Clean dataframe of Spotify tracks}
\usage{
clean_track_metadata(tracks_df)
}
\arguments{
\item{tracks_df}{dataframe returned by get_track_metadata()}
}
\value{
cleaned dataframe
}
\description{
This function unnests track artists and track names and formats the required artist, year, and track_name columns.
}
