% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tracks.R
\name{clean_tracks}
\alias{clean_tracks}
\title{Clean dataframe of Spotify tracks}
\usage{
clean_tracks(tracks_df)
}
\arguments{
\item{tracks_df}{dataframe returned by get_tracks()}
}
\value{
cleaned Dataframe with track info.
}
\description{
This function unnests track artists and track names and formats the required artist, year, and track_name columns.
}
