% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oo_composers.R
\name{get_oo_composers}
\alias{get_oo_composers}
\title{Get composer information from Open Opus API}
\usage{
get_oo_composers()
}
\value{
Dataframe with composer info.
}
\description{
This function returns a dataframe with the name, birth year, and death year of classical composers in the Open Opus database.
}
