% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_metadata.R
\name{get_track_metadata}
\alias{get_track_metadata}
\title{Get metadata on all tracks in Spotify playlist}
\usage{
get_track_metadata(playlist_id)
}
\arguments{
\item{playlist_id}{a \href{https://developer.spotify.com/documentation/web-api/concepts/spotify-uris-ids}{Spotify playlist ID}}
}
\value{
data.frame with raw track info.
}
\description{
This function returns a data.frame with the name, artists, and recording year of all tracks in a playlist.
}
