% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cards.R
\name{make_cards}
\alias{make_cards}
\title{Make playing cards.}
\usage{
make_cards(
  tracks,
  file_name = NULL,
  dir = "./",
  card_size = "small",
  paper_size = "a4",
  color = FALSE
)
}
\arguments{
\item{tracks}{data.frame containing the columns artist, year, and track_name}

\item{file_name}{a file name ending in .pdf to write the output to (default: NULL, set name automatically based on parameters)}

\item{dir}{an existing path to write the file (default: "./", the current working directory)}

\item{card_size}{"small" (3.8 cm) or "original" (6.5 cm) (default: "small")}

\item{paper_size}{size "a4" or "letter" (default: "a4")}

\item{color}{color font based on year (default: FALSE)}
}
\value{
PDF file
}
\description{
This function makes a pdf of cards with song info on one side and a QR code to the Spotify URL on the other side.
}
