% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo_composers.R
\docType{data}
\name{oo_composers}
\alias{oo_composers}
\title{Composer data from Open Opus}
\format{
\subsection{\code{oo_composers}}{

A data frame with the following columns:
\describe{
\item{complete_name}{Name of the composer}
\item{epoch}{Classical music era}
\item{birth}{Birth year of the composer}
\item{death}{Death year of the composer}
}
}
}
\source{
\href{https://github.com/openopus-org/openopus_api}{Open Opus API}
}
\usage{
oo_composers
}
\description{
Sample output of composer information generated by \code{\link[=get_oo_composers]{get_oo_composers()}}.
}
\keyword{datasets}
