% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spotify_credentials.R
\name{set_spotify_credentials}
\alias{set_spotify_credentials}
\title{Set Spotify API credentials}
\usage{
set_spotify_credentials(spotify_client_id, spotify_client_secret)
}
\arguments{
\item{spotify_client_id}{Spotify client ID}

\item{spotify_client_secret}{Spotify client secret}
}
\description{
This function checks the format of the supplied Spotify credentials and stories them in the environment variables SPOTIFY_CLIENT_ID and SPOTIFY_CLIENT_SECRET.
}
