% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_metadata.R
\docType{data}
\name{track_metadata}
\alias{track_metadata}
\title{Metadata of tracks in a Spotify playlist}
\format{
\subsection{\code{track_metadata}}{

A data frame with the following columns:
\describe{
\item{url}{External Spotify URL of track}
\item{artist}{Concatenated artist names}
\item{year}{Release year of track}
\item{track_name}{Name of track}
}
}
}
\source{
Spotify API
}
\usage{
track_metadata
}
\description{
Sample output of cleaned track metadata generated by \code{\link[=clean_track_metadata]{clean_track_metadata()}}.
}
\keyword{datasets}
