% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_metadata_raw.R
\docType{data}
\name{track_metadata_raw}
\alias{track_metadata_raw}
\title{Raw metadata of tracks in a Spotify playlist}
\format{
\subsection{\code{track_metadata_raw}}{

A data frame with 114 rows and 40 columns.
}
}
\source{
Spotify API
}
\usage{
track_metadata_raw
}
\description{
Sample output of raw track metadata generated by \code{\link[=get_track_metadata]{get_track_metadata()}}.
}
\keyword{datasets}
