% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_track_metadata.R
\name{clean_track_metadata}
\alias{clean_track_metadata}
\title{Clean dataframe of Spotify tracks}
\usage{
clean_track_metadata(tracks_df, artists_n = 10, artists_char = 34)
}
\arguments{
\item{tracks_df}{dataframe returned by \code{\link[=get_track_metadata]{get_track_metadata()}}}

\item{artists_n}{number of artists to include (default: 10)}

\item{artists_char}{cumulative number of characters to include in the artist field (default: 34)}
}
\value{
dataframe
}
\description{
This function unnests track artists and track names and formats the required artist, year, and track_name columns.
}
