% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{Get playlist tracks from Spotify API}
\usage{
get_tracks(playlist_id)
}
\arguments{
\item{playlist_id}{Spotify playlist ID string.}
}
\value{
Dataframe with track info.
}
\description{
This function returns a dataframe with the name, artists, and recording year of all tracks in a playlist.
}
\examples{
get_tracks("6i2Qd6OpeRBAzxfscNXeWp")
}
