% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cards.R
\name{make_cards}
\alias{make_cards}
\title{Make playing cards.}
\usage{
make_cards(
  tracks,
  file = NULL,
  card_size = "small",
  paper_size = "a4",
  color = FALSE
)
}
\arguments{
\item{tracks}{dataframe containing the columns artist, year, and track_name}

\item{file}{a valid path to a pdf file (default: NULL, save file in output/)}

\item{card_size}{"small" (3.8 cm) or "original" (6.5 cm) (default: "small")}

\item{paper_size}{size "a4" or "letter" (default: "a4")}

\item{color}{color font based on year (default: FALSE)}
}
\value{
PDFs saved to /output.
}
\description{
This function makes a pdf of cards with song info on one side and a QR code to the Spotify URL on the other side.
}
