% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_metadata.R
\docType{data}
\name{track_metadata}
\alias{track_metadata}
\title{Metadata of tracks in a Spotify playlist}
\format{
\subsection{\code{track_metadata}}{

A data frame with 114 rows and 5 columns:
\describe{
\item{artist1, artist}{First artist & first two artists concatenated}
\item{year}{Release year of track}
\item{track_name}{Name of track}
\item{url}{External Spotify URL of track}
}
}
}
\source{
Spotify API
}
\usage{
track_metadata
}
\description{
Sample output of cleaned track metadata generated by `clean_track_metadata().
}
\keyword{datasets}
